requirements.txt
numpy
flask
flask_cors
requests
pandas
mysql-connector-python